Fixed circular crop overflow bug (#179).

+++ src/mapcraftercore/mc/worldcrop.cpp
@@ -210,7 +210,7 @@ void WorldCrop::setCenter(const BlockPos& pos) {
 	type = CIRCULAR;
 }
 
-void WorldCrop::setRadius(int radius) {
+void WorldCrop::setRadius(long radius) {
 	this->radius = radius;
 	type = CIRCULAR;
 }
@@ -225,8 +225,8 @@ bool WorldCrop::isRegionContained(const mc::RegionPos& region) const {
 		// check roughly whether at least one block of the region is included
 		// use the midpoint of the region and determine the distance to the center
 		BlockPos region_center(region.x * 512 + 256, region.z * 512 + 256, 0);
-		int dx = region_center.x - center.x;
-		int dz = region_center.z - center.z;
+		long dx = region_center.x - center.x;
+		long dz = region_center.z - center.z;
 		// and check whether it is at most radius + size of a region blocks away
 		return (radius+512)*(radius+512) >= dx*dx + dz*dz;
 	}
@@ -243,8 +243,8 @@ bool WorldCrop::isChunkContained(const mc::ChunkPos& chunk) const {
 		// circular crop:
 		// do the same thing doing with regions
 		BlockPos chunk_center(chunk.x * 16 + 8, chunk.z * 16 + 8, 0);
-		int dx = chunk_center.x - center.x;
-		int dz = chunk_center.z - center.z;
+		long dx = chunk_center.x - center.x;
+		long dz = chunk_center.z - center.z;
 		return (radius+16)*(radius+16) >= dx*dx + dz*dz;
 	}
 
@@ -267,8 +267,8 @@ bool WorldCrop::isBlockContainedXZ(const mc::BlockPos& block) const {
 	} else if (type == CIRCULAR) {
 		// circular crop:
 		// also check distance to the center (like with regions and chunks)
-		int dx = block.x - center.x;
-		int dz = block.z - center.z;
+		long dx = block.x - center.x;
+		long dz = block.z - center.z;
 		return radius*radius >= dx*dx + dz*dz;
 	}
